@@ -44,7 +44,7 @@ that could potentially be improved'),
     zip_safe=False,
     install_requires=[
         'pyyaml>5.4',
-        'aws-sam-translator>=1.82.0',
+        'aws-sam-translator>=1.87.0',
         'jsonpatch',
         'jsonschema>=3.0,<5',
         'networkx>=2.4,<4',
@@ -54,7 +54,7 @@ that could potentially be improved'),
         'sympy>=1.0.0',
         'regex>=2021.7.1',
     ],
-    python_requires='>=3.7, <=4.0, !=4.0',
+    python_requires='>=3.8, <=4.0, !=4.0',
     entry_points={
         'console_scripts': [
             'cfn-lint = cfnlint.__main__:main'
@@ -69,7 +69,6 @@ that could potentially be improved'),
         'Natural Language :: English',
         'Operating System :: OS Independent',
         'Programming Language :: Python :: 3',
-        'Programming Language :: Python :: 3.7',
         'Programming Language :: Python :: 3.8',
         'Programming Language :: Python :: 3.9',
         'Programming Language :: Python :: 3.10',
