@@ -2,6 +2,7 @@
 Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 SPDX-License-Identifier: MIT-0
 """
+
 from cfnlint.rules import CloudFormationLintRule, RuleMatch
 
 
@@ -25,8 +26,10 @@ class Configuration(CloudFormationLintRule):
     def match(self, cfn):
         matches = []
 
-        conditions = cfn.template.get("Conditions", {})
-        if conditions:
+        if "Conditions" not in cfn.template:
+            return matches
+        conditions = cfn.template.get("Conditions", None)
+        if isinstance(conditions, dict):
             for condname, condobj in conditions.items():
                 if not isinstance(condobj, dict):
                     message = "Condition {0} has invalid property"
@@ -51,5 +54,12 @@ class Configuration(CloudFormationLintRule):
                                         message.format(condname, k),
                                     )
                                 )
+        else:
+            matches.append(
+                RuleMatch(
+                    ["Conditions"],
+                    "Condition must be an object",
+                )
+            )
 
         return matches
