@@ -2,6 +2,7 @@
 Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 SPDX-License-Identifier: MIT-0
 """
+
 from test.unit.rules import BaseRuleTestCase
 
 from cfnlint.rules.resources.iam.Policy import Policy  # pylint: disable=E0401
@@ -26,11 +27,11 @@ class TestPropertyIamPolicies(BaseRuleTestCase):
     def test_file_negative(self):
         """Test failure"""
         self.helper_file_negative(
-            "test/fixtures/templates/bad/resources/iam/iam_policy.yaml", 12
+            "test/fixtures/templates/bad/resources/iam/iam_policy.yaml", 16
         )
 
     def test_file_resource_negative(self):
         """Test failure"""
         self.helper_file_negative(
-            "test/fixtures/templates/bad/resources/iam/resource_policy.yaml", 4
+            "test/fixtures/templates/bad/resources/iam/resource_policy.yaml", 6
         )
