@@ -2,6 +2,7 @@
 Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 SPDX-License-Identifier: MIT-0
 """
+
 from test.unit.rules import BaseRuleTestCase
 
 from cfnlint.rules.resources.properties.AvailabilityZone import (
@@ -17,7 +18,8 @@ class TestPropertyAvailabilityZone(BaseRuleTestCase):
         super(TestPropertyAvailabilityZone, self).setUp()
         self.collection.register(AvailabilityZone())
         self.success_templates = [
-            "test/fixtures/templates/good/resources/properties/az.yaml"
+            "test/fixtures/templates/good/resources/properties/az.yaml",
+            "test/fixtures/templates/good/resources/properties/az_cdk.yaml",
         ]
 
     def test_file_positive(self):
