@@ -2,6 +2,7 @@
 Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 SPDX-License-Identifier: MIT-0
 """
+
 from test.unit.rules import BaseRuleTestCase
 
 from cfnlint.rules.resources.properties.PropertiesTemplated import (
@@ -17,7 +18,8 @@ class TestPropertiesTemplated(BaseRuleTestCase):
         super(TestPropertiesTemplated, self).setUp()
         self.collection.register(PropertiesTemplated())
         self.success_templates = [
-            "test/fixtures/templates/good/resources/properties/templated_code.yaml"
+            "test/fixtures/templates/good/resources/properties/templated_code.yaml",
+            "test/fixtures/templates/good/resources/properties/templated_code_sam.yaml",
         ]
 
     def test_file_positive(self):
