@@ -2,6 +2,7 @@
 Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 SPDX-License-Identifier: MIT-0
 """
+
 import json
 from test.unit.rules import BaseRuleTestCase
 
@@ -29,7 +30,7 @@ class TestResourceConfiguration(BaseRuleTestCase):
     def test_file_negative(self):
         """Test failure"""
         self.helper_file_negative(
-            "test/fixtures/templates/bad/properties_required.yaml", 11
+            "test/fixtures/templates/bad/properties_required.yaml", 12
         )
 
     def test_file_negative_generic(self):
@@ -56,5 +57,5 @@ class TestSpecifiedCustomResourceRequiredProperties(TestResourceConfiguration):
         """Test failure"""
         # Additional Custom::SpecifiedCustomResource failure detected with custom spec
         self.helper_file_negative(
-            "test/fixtures/templates/bad/properties_required.yaml", 12
+            "test/fixtures/templates/bad/properties_required.yaml", 13
         )
